requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
albucore==0.0.23
albumentations==1.4.24
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.7.2
anyio==4.13.0
attrs==26.1.0
blis==1.3.3
catalogue==2.0.10
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
cloudpathlib==0.23.0
cnocr==2.3.2.3
cnstd==1.2.7.1
colorama==0.4.6
colorlog==6.10.1
confection==1.3.3
contourpy==1.3.3
cycler==0.12.1
cymem==2.0.13
doclayout_yolo==0.0.4
einops==0.8.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
entmax==1.3
fastapi==0.135.2
filelock==3.25.2
flatbuffers==25.12.19
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.3.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
Jinja2==3.1.6
kiwisolver==1.5.0
latex2sympy2==1.9.1
lightning-utilities==0.15.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
multidict==6.7.1
munch==4.0.0
murmurhash==1.0.15
networkx==3.6.1
numpy==2.4.4
omegaconf==2.3.0
onnx==1.21.0
onnxruntime==1.24.4
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
optimum==2.1.0
optimum-onnx==0.1.0
packaging==26.0
pandas==3.0.2
pillow==12.2.0
pix2tex==0.1.4
pix2text==1.1.6
platformdirs==4.9.4
polars==1.39.3
polars-runtime-32==1.39.3
preshed==3.0.13
propcache==0.4.1
protobuf==6.33.6
psutil==7.2.2
py-cpuinfo==9.0.0
pyclipper==1.4.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
PyMuPDF==1.27.2.2
pynput==1.8.1
pyparsing==3.3.2
PyQt6==6.11.0
PyQt6-Qt6==6.11.0
PyQt6-WebEngine==6.11.0
PyQt6-WebEngine-Qt6==6.11.0
PyQt6_sip==13.11.1
pyreadline3==3.5.4
PySide6==6.11.0
PySide6_Addons==6.11.0
PySide6_Essentials==6.11.0
pyspellchecker==0.9.0
python-dateutil==2.9.0.post0
python-multipart==0.0.22
pytorch-lightning==2.6.1
PyYAML==6.0.3
rapidocr==3.7.0
regex==2026.4.4
requests==2.33.1
rich==14.3.3
safetensors==0.7.0
scipy==1.17.1
screeninfo==0.8.1
seaborn==0.13.2
sentry-sdk==2.57.0
setuptools==81.0.0
shapely==2.1.2
shellingham==1.5.4
shiboken6==6.11.0
simsimd==6.5.16
six==1.17.0
smart_open==7.5.1
smmap==5.0.3
spacy==3.8.13
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.3
starlette==1.0.0
stringzilla==4.6.0
sympy==1.14.0
thinc==8.3.13
thop==0.1.1.post2209072238
timm==0.5.4
tokenizers==0.22.2
torch==2.11.0
torchmetrics==1.9.0
torchvision==0.26.0
tqdm==4.67.3
transformers==4.57.6
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
ultralytics==8.4.34
ultralytics-thop==2.0.18
Unidecode==1.4.0
urllib3==2.6.3
uvicorn==0.42.0
wandb==0.25.1
wasabi==1.1.3
weasel==1.0.0
wheel==0.46.3
wrapt==2.1.2
x-transformers==0.15.0
yarl==1.23.0
